Remove polarization default, fix prograde filter
The "polarization" keyword no longer has a default argument.  This was
mainly to aid debugging the prograde filtering, in case the default was
being applied in cases where prograde polarization should have been
propagated through the function cascade.

A bug in the prograde filtering was fixed.  Previously,
filtering:get_filter grouped prograde and linear polarized behaviour
together.  This was a misunderstanding of the Meza-Fajardo paper
equations 27 & 28.